Non-connector connects should not delete connectors

+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/AbstractConnectOperator.java
@@ -246,8 +246,10 @@ public abstract class AbstractConnectOperator<C extends KubernetesClient, T exte
 
     private Future<Void> reconcileConnector(Reconciliation reconciliation, String host, KafkaConnectApi apiClient, boolean useResources, String connectorName, KafkaConnector connector) {
         if (connector == null) {
-            log.debug("{}: {}} cluster: deleting connector: {}", reconciliation, kind(), connectorName);
-            return apiClient.delete(host, KafkaConnectCluster.REST_API_PORT, connectorName);
+            if (useResources) {
+                log.debug("{}: {}} cluster: deleting connector: {}", reconciliation, kind(), connectorName);
+                return apiClient.delete(host, KafkaConnectCluster.REST_API_PORT, connectorName);
+            }
         } else {
             log.debug("{}: {}} cluster: creating/updating connector: {}", reconciliation, kind(), connectorName);
             if (connector.getSpec() == null) {
